backend/requirements.txt
flask==2.2.5
flask-cors==4.0.0
sqlalchemy>=2.0.36
python-dotenv==1.0.0
pymysql==1.1.0
cryptography==41.0.7
marshmallow==3.20.1
gunicorn==21.2.0
azure-storage-blob==12.19.0
azure-storage-file-datalake==12.14.0
azure-storage-file-share==12.19.0
azure-storage-queue==12.8.0
azure-data-tables==12.4.0
azure-identity==1.15.0
azure-core==1.30.0
azure-ai-textanalytics==5.3.0
pyarrow>=18.0.0
pandas>=2.2.0
sqlglot==24.0.0
fastavro>=1.11.0
openpyxl==3.1.2
xmltodict==0.13.0
requests==2.31.0
# Oracle Database support - use one of the following:
# oracledb>=1.4.0  # Recommended: Official Oracle driver (newer)
# cx_Oracle>=8.3.0  # Alternative: Legacy Oracle driver

